requirements.txt
# Flask Backend for MWECAU ICT Club Website

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7

# Environment Variables
python-dotenv==1.0.0

# Date/Time
pytz==2023.3
python-dateutil==2.8.2

# HTTP
requests==2.31.0
